Reorganize files to support multiple pages
