+++ packages/experiments/src/pages/index/html.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 
-const criticalCSS = fs.readFileSync(__dirname + '/../css/critical.css', 'utf8');
+const criticalCSS = fs.readFileSync(__dirname + '/../../css/critical.css', 'utf8');
 
 export const html = `
   <!doctype html>
