+++ packages/experiments/src/pages/index/hydrateEditor.ts
@@ -1,0 +1,5 @@
+export const hydrateEditor = view => {
+  const editorDiv = document.querySelector('#editor');
+  editorDiv.innerHTML = '';
+  editorDiv.appendChild(view.dom)
+};
